Changed location of one-time witness checking function
