+++ crates/sui-framework/src/natives/object.rs
@@ -3,7 +3,7 @@
 
 use crate::EventType;
 use move_binary_format::errors::PartialVMResult;
-use move_core_types::{account_address::AccountAddress, language_storage::TypeTag};
+use move_core_types::account_address::AccountAddress;
 use move_vm_runtime::native_functions::NativeContext;
 use move_vm_types::{
     gas_schedule::NativeCostIndex,
@@ -49,32 +49,6 @@ pub fn borrow_uid(
     let cost = native_gas(context.cost_table(), NativeCostIndex::SIGNER_BORROW, 0);
 
     Ok(NativeResult::ok(cost, smallvec![id_field]))
-}
-
-pub fn is_one_time_witness(
-    context: &mut NativeContext,
-    mut ty_args: Vec<Type>,
-    args: VecDeque<Value>,
-) -> PartialVMResult<NativeResult> {
-    debug_assert!(ty_args.len() == 1);
-    debug_assert!(args.len() == 1);
-
-    // unwrap safe because the interface of native function guarantees it.
-    let type_tag = context.type_to_type_tag(&ty_args.pop().unwrap())?;
-
-    // TODO: what should the cost of this be?
-    let cost = native_gas(context.cost_table(), NativeCostIndex::LENGTH, 1);
-
-    // If a struct type has the same name as the module that defines it but capitalized, it means
-    // that it's a characteristic type (which is one way of implementing a one-time witness
-    // type). This is checked in the char_type validator pass (a type with this type of name that
-    // does not have all properties required of a characteristic type will cause a validator error).
-    Ok(NativeResult::ok(
-        cost,
-        smallvec![Value::bool(
-            matches!(type_tag, TypeTag::Struct(struct_tag) if struct_tag.name.to_string() == struct_tag.module.to_string().to_ascii_uppercase())
-        )],
-    ))
 }
 
 pub fn delete_impl(
